Rename Notificaitons component file
